Update essential-cardano-list.md
A new NFT marketplace (exnft.io) is being built on both Cardano and Ergo blockchains. It will allow users to auction and trade Cardano and Ergo NFTs. It would be awesome if the site was added to the list. You can find more information on our twitter:  https://twitter.com/exnftsupport