Added gift card generator
RNDfr's Gift Card Generator lets you generate fancy paper wallets bundled with instructions to retrieve its funds. You can download it as a PDF or print it directly. You can use the template and write your words or let the app generate new mnemonics for you. The app is available in English and French and is hosted by a reliable stake pool.
https://randomfrenchpool.com/en/cards