Rebranding to AdaMuesliSwap
There has been considerable confusion regarding the co-existance of the two DEX projects MuesliSwap for SmartBCH (https://bch.muesliswap.com) and MuesliSwap for Cardano (https://ada.muesliswap.com). Therefore, the DEX will from now on be using the name "AdaMuesliSwap" when referring to the Cardano branch of the DEX project.
This rebranding should also be reflected in this list.