Add NFTJam.io
We're building a NFT Marketplace for Cardano with metadata search/filtering feature, and giving projects their own market space to display their NFT Collections. We are on track to implement Smart Contract trade that would also let yet-to-support SC wallets like Daedalus to participate. We hope to be able to join this list.